In the DLT_LINUX_SLL2 header, the hardware address type is 2 octets.
(cherry picked from commit a1bc985bb6aac4522b6bea725a2ad7e65d6cea0f)

+++ pcap/sll.h
@@ -99,7 +99,7 @@ struct sll2_header {
 	uint16_t sll2_protocol;			/* protocol */
 	uint16_t sll2_reserved_mbz;		/* reserved - must be zero */
 	uint32_t sll2_if_index;			/* 1-based interface index */
-	uint8_t  sll2_hatype;			/* link-layer address type */
+	uint16_t sll2_hatype;			/* link-layer address type */
 	uint8_t  sll2_pkttype;			/* packet type */
 	uint8_t  sll2_halen;			/* link-layer address length */
 	uint8_t  sll2_addr[SLL_ADDRLEN];	/* link-layer address */
